requirements.txt
flask>=3.0
PyMuPDF>=1.24
openai>=1.0
google-api-python-client>=2.0
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.0
werkzeug>=3.0
